Update - Timesheet Change TypeValue

+++ Bovis.Common.Model/Tables/TB_Timesheet_Otro.cs
@@ -8,8 +8,8 @@ namespace Bovis.Common.Model.Tables
         [Column("IdTs_Otro"), NotNull, PrimaryKey, Identity] public int IdTimesheet_Otro { get; set; }
         [Column("NukidTs"), NotNull] public int IdTimeSheet { get; set; }
         [Column("Chdescripcion"), NotNull] public string Descripcion { get; set; }
-        [Column("Numdias"), NotNull] public decimal Dias { get; set; }
-        [Column("NumTdedicacion"), NotNull] public decimal TDedicacion { get; set; }
+        [Column("Numdias"), NotNull] public float Dias { get; set; }
+        [Column("NumTdedicacion"), NotNull] public float TDedicacion { get; set; }
         [Column("Boactivo"), NotNull] public bool Activo { get; set; }
     }
 }
